+++ apps/mis-web/src/models/UserSchemaModel.ts
@@ -12,7 +12,8 @@
 
 import { Static, Type } from "@sinclair/typebox";
 
-import { ClusterAccountInfo_ImportStatus, PlatformRole, TenantRole, UserRole, UserStatus } from "./User";
+import { AccountState, ClusterAccountInfo_ImportStatus, DisplayedAccountState, DisplayedUserState, PlatformRole,
+  TenantRole, UserRole, UserStateInAccount, UserStatus } from "./User";
 
 // 这个Model重新用typebox定义了
 // 定义Schema时无法复用的@scow/protos/build/server中的interface
@@ -90,6 +91,9 @@ export const Account = Type.Object({
   accountName: Type.String(),
   userCount: Type.Number(),
   blocked: Type.Boolean(),
+  state: Type.Optional(Type.Enum(AccountState)),
+  displayedState: Type.Optional(Type.Enum(DisplayedAccountState)),
+  isInWhitelist: Type.Optional(Type.Boolean()),
   ownerId: Type.String(),
   ownerName: Type.String(),
   comment: Type.String(),
@@ -138,6 +142,8 @@ export const AccountUserInfo = Type.Object({
   storageQuotas: Type.Record(Type.String(), Type.Number()),
   jobChargeLimit:Type.Optional(Money),
   usedJobChargeLimit: Type.Optional(Money),
+  userStateInAccount: Type.Optional(Type.Enum(UserStateInAccount)),
+  displayedUserState: Type.Optional(Type.Enum(DisplayedUserState)),
 });
 export type AccountUserInfo = Static<typeof AccountUserInfo>;