+++ libs/web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scow/lib-web
 
+## 1.3.1
+
+### Patch Changes
+
+- 146e19f: 去掉导航栏多余的下划线
+- 850a7ee: 修改 UserAccount 实体中原 status 字段名为 blocked_in_cluster ,表示在集群中是否为封锁状态
+  增加字段 state ,允许写入的值为 "NORMAL" , "BLOCKED_BY_ADMIN" 的枚举值
+  页面增加用户在账户下的 限额 的状态的显示
+
 ## 1.3.0
 
 ### Minor Changes